--- .gitignore
@@ -1,4 +1,0 @@
-bool_true = True
-bool_false = False
-print(type(bool_true))
-
