+++ test.py
@@ -1,2 +1,1 @@
-print("Hello world")
-print("New date")
+word =
